Must not apply labels to nested objects
